+++ package.json
@@ -18,6 +18,7 @@
     "jsonwebtoken": "^8.5.1",
     "pg": "^8.7.3",
     "reflect-metadata": "^0.1.13",
+    "supertest": "^6.3.3",
     "swagger-ui-express": "^4.3.0",
     "tsyringe": "^4.6.0",
     "typeorm": "0.2.45"
@@ -28,6 +29,7 @@
     "@types/jest": "^28.1.6",
     "@types/jsonwebtoken": "^8.5.8",
     "@types/multer": "^1.4.7",
+    "@types/supertest": "^2.0.12",
     "@types/swagger-ui-express": "^4.1.3",
     "@types/uuid": "^8.3.3",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
