+++ src/modules/cars/useCases/createCategory/CreateCategoryController.spec.ts
@@ -1,0 +1,8 @@
+import request from "supertest";
+import { app } from "../../../../shared/infra/http/app";
+
+describe("Create Category Controller", () => {
+    it("hello world!", async()=>{
+        await request(app).get("/cars/available").expect(200);
+    })
+})
